fix(vmaas-go): recover from panic and respond 500
