+++ vmaas-go/webapp/controllers/vulnerabilities.go
@@ -6,6 +6,7 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/redhatinsights/vmaas-lib/vmaas"
 	"github.com/redhatinsights/vmaas/base/core"
+	"github.com/redhatinsights/vmaas/base/utils"
 )
 
 func VulnerabilitiesHandler(c *gin.Context) {
@@ -17,7 +18,7 @@ func VulnerabilitiesHandler(c *gin.Context) {
 
 	vulnerabilities, err := core.VmaasAPI.Vulnerabilities(&request)
 	if err != nil {
-		LogAndRespError(c, err)
+		utils.LogAndRespError(c, err)
 		return
 	}
 	c.JSON(http.StatusOK, vulnerabilities)
@@ -29,14 +30,14 @@ func VulnerabilitiesPostHandler(c *gin.Context) {
 	}
 	request, err := bindValidateJSON(c)
 	if err != nil {
-		LogAndRespBadRequest(c, err)
+		utils.LogAndRespBadRequest(c, err)
 		return
 	}
 
 	if request.Extended {
 		vulnerabilities, err := core.VmaasAPI.VulnerabilitiesExtended(request)
 		if err != nil {
-			LogAndRespError(c, err)
+			utils.LogAndRespError(c, err)
 			return
 		}
 		c.JSON(http.StatusOK, vulnerabilities)
@@ -45,7 +46,7 @@ func VulnerabilitiesPostHandler(c *gin.Context) {
 
 	vulnerabilities, err := core.VmaasAPI.Vulnerabilities(request)
 	if err != nil {
-		LogAndRespError(c, err)
+		utils.LogAndRespError(c, err)
 		return
 	}
 	c.JSON(http.StatusOK, vulnerabilities)
